+++ src/white.cpp
@@ -3,8 +3,8 @@
 #include "FastNoise.h"
 
 [[cpp11::register]]
-cpp11::writable::doubles_matrix white_2d_c(int height, int width, int seed, double freq, int pertube, double pertube_amp) {
-  cpp11::writable::doubles_matrix noise(height, width);
+cpp11::writable::doubles_matrix<> white_2d_c(int height, int width, int seed, double freq, int pertube, double pertube_amp) {
+  cpp11::writable::doubles_matrix<> noise(height, width);
   int i,j;
   double new_i, new_j;
   FastNoise noise_gen;
@@ -30,8 +30,8 @@ cpp11::writable::doubles_matrix white_2d_c(int height, int width, int seed, doub
 }
 
 [[cpp11::register]]
-cpp11::writable::doubles_matrix white_3d_c(int height, int width, int depth, int seed, double freq, int pertube, double pertube_amp) {
-  cpp11::writable::doubles_matrix noise(height, width * depth);
+cpp11::writable::doubles_matrix<> white_3d_c(int height, int width, int depth, int seed, double freq, int pertube, double pertube_amp) {
+  cpp11::writable::doubles_matrix<> noise(height, width * depth);
   int i,j,k;
   double new_i, new_j, new_k;
   FastNoise noise_gen;
@@ -60,8 +60,8 @@ cpp11::writable::doubles_matrix white_3d_c(int height, int width, int depth, int
 }
 
 [[cpp11::register]]
-cpp11::writable::doubles_matrix white_4d_c(int height, int width, int depth, int time, int seed, double freq, int pertube, double pertube_amp) {
-  cpp11::writable::doubles_matrix noise(height, width * depth * time);
+cpp11::writable::doubles_matrix<> white_4d_c(int height, int width, int depth, int time, int seed, double freq, int pertube, double pertube_amp) {
+  cpp11::writable::doubles_matrix<> noise(height, width * depth * time);
   int i,j,k,l;
   double new_i, new_j, new_k, new_l;
   FastNoise noise_gen;